pyproject.toml
[tool.poetry]
name = "sprite-splitter"
version = "0.1.0"
description = ""
authors = ["jkarenko <juho.karenko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.3.0"
wand = "^0.6.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
